[Process-Swap-On-Navigation] WebKit hangs when going back to a form submission's page due to Process-Swap-On-Navigation on iOS 12.2 and higher https://bugs.webkit.org/show_bug.cgi?id=197097 <rdar://problem/50048318>
Reviewed by Alex Christensen.

Source/WebCore:

* loader/EmptyFrameLoaderClient.h:
* loader/FrameLoader.cpp:
(WebCore::FrameLoader::checkLoadCompleteForThisFrame):
If we still have a provisional item (new load has been started or we did not clear it because we're about to retry),
then let the client know the load will continue by passing an extra flag with the didFailProvisionalLoad.

* loader/FrameLoaderClient.h:
* loader/FrameLoaderTypes.h:

Source/WebKit:

The issue is that when doing a history navigation with form data, we try to resubmit the form.
We initially use the ReturnCacheDataDontLoad cache policy without prompting the user to get
cached data. If this fails, we sent a DidFailProvisionalLoadForFrame IPC to the UIProcess but
the WebContent process actually retries the load with ReturnCacheDataElseLoad cache policy.
This new load triggers a new decidePolicyForNavigationAction.

This would cause trouble with process-swapping because we normally destroy the
ProvisionalPageProxy as soon as we receive a DidFailProvisionalLoadForFrame IPC from the
provisional process. As a result, the provisional process would not be able to retry.

To address the issue, we add pass a flag with the DidFailProvisionalLoadForFrame IPC to
if the load will continue or not. When this flag is set, the UIProcess does not destroy
the provisional page.

* Scripts/webkit/messages.py:
Minor IPC message handler build fix.

* UIProcess/API/Cocoa/WKURLSchemeTask.mm:
(-[WKURLSchemeTaskImpl _onlyIfCached]):
* UIProcess/API/Cocoa/WKURLSchemeTaskPrivate.h:
Add new _onlyIfCached SPI to WKURLSchemeTaskPrivate to allow for API testing.

* UIProcess/ProvisionalPageProxy.cpp:
(WebKit::ProvisionalPageProxy::cancel):
Pass WillContinueLoading flag.

(WebKit::ProvisionalPageProxy::validateInput):
Consider IPC with a navigationID of 0 as valid. When the navigation is triggered by the
WebContent process, it sends us a DecidePolicyForNavigationActionAsync IPC with a navigationID
of 0, until the UIProcess can generate a valid identifier.

(WebKit::ProvisionalPageProxy::didFailProvisionalLoadForFrame):

(WebKit::ProvisionalPageProxy::decidePolicyForNavigationActionAsync):
When the WebPageProxy has generated a navigationID for the new WebContent process-initiated
navigation, update m_navigationID so that follow-up IPC is considered valid.

* UIProcess/ProvisionalPageProxy.h:
* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didFailProvisionalLoadForFrame):
(WebKit::WebPageProxy::didFailProvisionalLoadForFrameShared):
(WebKit::WebPageProxy::decidePolicyForNavigationActionAsyncShared):
(WebKit::WebPageProxy::decidePolicyForNavigationAction):
* UIProcess/WebPageProxy.h:
* UIProcess/WebPageProxy.messages.in:
* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::dispatchDidFailProvisionalLoad):
* WebProcess/WebCoreSupport/WebFrameLoaderClient.h:

Source/WebKitLegacy/mac:

* WebCoreSupport/WebFrameLoaderClient.h:
* WebCoreSupport/WebFrameLoaderClient.mm:
(WebFrameLoaderClient::dispatchDidFailProvisionalLoad):

Source/WebKitLegacy/win:

* WebCoreSupport/WebFrameLoaderClient.cpp:
(WebFrameLoaderClient::dispatchDidFailProvisionalLoad):
* WebCoreSupport/WebFrameLoaderClient.h:

Tools:

Add API test coverage.

* TestWebKitAPI/Tests/WebKitCocoa/ProcessSwapOnNavigation.mm:
(-[PSONScheme webView:startURLSchemeTask:]):

Identifier: 211400@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244540 268f45cc-cd09-0410-ab3c-d52691b4dbfc