WebCore::StyleColorScheme should not have explicitly-declared copy constructor https://bugs.webkit.org/show_bug.cgi?id=197412
Reviewed by Don Olmstead.

Either we need to explicitly declare a copy assignment operator here, or the copy
constructor needs to be removed. Having one without the other causes a huge warning spam
with GCC 9. In this case, the copy constructor is redundant because it's identical to an
implicitly-declared copy constructor, so let's just remove it.

* rendering/style/StyleColorScheme.h:


Identifier: 211593@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244768 268f45cc-cd09-0410-ab3c-d52691b4dbfc