Temporarily disable generate-xcfilelists https://bugs.webkit.org/show_bug.cgi?id=197619 <rdar://problem/50507392>
Reviewed by Alex Christensen.

We need to perform a significant update to the generate-xcfilelist
scripts. This work involves coordinated work with another facility. If
the work does not occur in tandem, the build will be broken. To avoid
this, disable the invoking of the scripts during the transition. The
checking will be restored once the new scripts are in place.

Source/JavaScriptCore:

* Scripts/check-xcfilelists.sh:

Source/WebCore:

No new tests -- no change in user-visible functionality.

* Scripts/check-xcfilelists.sh:

Source/WebKit:

* Scripts/check-xcfilelists.sh:

Tools:

* DumpRenderTree/Scripts/check-xcfilelists.sh:
* WebKitTestRunner/Scripts/check-xcfilelists.sh:


Identifier: 211772@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@244987 268f45cc-cd09-0410-ab3c-d52691b4dbfc