Add option to build-webkit to control whether or not XCBuild is used https://bugs.webkit.org/show_bug.cgi?id=197668 <rdar://problem/50549728>
Reviewed by Tim Horton.

Add --[no-]xcbuild to build-webkit to force the use of XCBuild or not.
Also update build-webkit and the makefiles with the foundation for
automatically using XCBuild when the conditions allow it. This latter
facility is currently turned off until Xcode fully supports building
WebKit with XCBuild.

.:

* Makefile.shared:

Tools:

* Scripts/build-webkit:
* Scripts/webkitdirs.pm:
(XcodeOptions):
(canUseXCBuild):


Identifier: 211816@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245037 268f45cc-cd09-0410-ab3c-d52691b4dbfc