[Legacy WebKit] REGRESSION (r238078): Crash in hardwareKeyboardAvailabilityChangedCallback() https://bugs.webkit.org/show_bug.cgi?id=197724 <rdar://problem/49725959>
Reviewed by Tim Horton.

Speculative fix for race condition. Between the time we receive a kGSEventHardwareKeyboardAvailabilityChangedNotification
notification and when we schedule execution to run on the WebThread the WebView that notification
was for may no longer exist. Take out a weak ptr on the WebView when we receive the notification
and check that we still have it once we are running on the WebThread.

* WebView/WebHTMLView.mm:
(hardwareKeyboardAvailabilityChangedCallback):

Identifier: 211855@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245076 268f45cc-cd09-0410-ab3c-d52691b4dbfc