[WHLSL] parseEffectfulAssignment should not call parseCallExpression directly https://bugs.webkit.org/show_bug.cgi?id=197890
Reviewed by Myles Maxfield.

callExpression already appears in effSuffix which is in effPrefix which is in effAssignment, so having it directly in effAssignment as well is useless (and ambiguous).
I've already fixed the grammar (https://github.com/gpuweb/WHLSL/commit/a07005f4d692fe3370618dca5db218992b362049), the grammar was always good, this patch is fixing the parser.

* Modules/webgpu/WHLSL/WHLSLParser.cpp:
(WebCore::WHLSL::Parser::parseEffectfulAssignment):


Identifier: 212023@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245302 268f45cc-cd09-0410-ab3c-d52691b4dbfc