inputmode="numeric" should show a number pad with digits 0-9, instead of the numeric keyplane https://bugs.webkit.org/show_bug.cgi?id=197916 <rdar://problem/50815427>
Reviewed by Timothy Hatcher.

Source/WebKit:

Use UIKeyboardTypeNumberPad instead of UIKeyboardTypeNumbersAndPunctuation when presenting a keyboard for a
field with inputmode="numeric". While the WhatWG specification merely requires the UA to display a keyboard
"capable of numeric input", it suggests that the keyboard should be "useful for PIN entry", which loosely
implies a number pad.

This is also generally in line with feedback from web developers.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView textInputTraits]):

Tools:

Re-enable this previously flaky test, and rebaseline the result to to expect UIKeyboardTypeNumberPad for
inputmode="numeric".

* TestWebKitAPI/Tests/ios/KeyboardInputTestsIOS.mm:
(TestWebKitAPI::TEST):


Identifier: 212047@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@245338 268f45cc-cd09-0410-ab3c-d52691b4dbfc